fix: remove dependency cycle by moving getDefaultContributionType into productType
